+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.3: Fix search error when using kinship and relative position
 
 * Fix incorrect UiObject2 being found when using waitForExists and waitUntilGone on 2 or more layers of BySelector
+* Fix Selector being overwritten when creating an instance of child UiObject2
 
 ## 1.0.2: Fix NullPointerException when calling UiObject2#getParent()